TA-29 Update deployment method
The previous method of deployment involved making manual changes to have
the correct project settings.  In order to make reliable deployments we
need a better process.  This change means that you can deploy to the
different environments with a one line command.
firebase dependency updated as the deploy command was failing.